requirements.txt
streamlit==1.26.0
google-generativeai==0.1.0
transformers==4.31.0
torch==2.0.1
faiss-cpu==1.7.4
numpy==1.24.2
sentence-transformers==2.2.2
